+++ includes/core/Enqueue.php
@@ -2,11 +2,11 @@
 
 namespace WP_Vue_Theme\core;
 
-class Enqueue
+class Enqueue extends \WP_Vue_Theme\wp_rest_api\BaseControler
 {
   public function init(){
-    
-    
+
+
     add_action( 'wp_enqueue_scripts', array( $this, 'enqueue' ) );
   }
   public function enqueue(){
@@ -14,14 +14,16 @@ class Enqueue
 
     wp_enqueue_script('app-js');
   }
-  
+
   public function register(){
     wp_register_script('app-js', get_template_directory_uri() . '/dist/main.js', array(), '1.0.0', true);
-    
+
+
     wp_localize_script('app-js', 'wpDataSite', array(
-      'siteUrl' => (string)get_site_url(),
-      'siteName' => (string)get_bloginfo('name'),
+      'siteUrl'         => (string)get_site_url(),
+      'siteName'        => (string)get_bloginfo('name'),
       'siteDescription' => (string)get_bloginfo('description'),
+      'themeEndPoint'   => $this->themeEndPoint
     ));
   }
 }